Poster: Move rating and status btns own components
